cffdump: add --once mode
Add an option to dump cmdstream only once, for example in cases where it
is executed once per tile, but seeing the decode for each tile is not
useful (ie. looking at gpu state that doesn't change per tile).  This
takes into account the execution mode (enabled_mask, ie BINNING vs GMEM
vs SYSMEM), so that on a6xx+ where the same cmdstream is used both for
BINNING and per-tile for GMEM, the cmdstream will be decoded both for
the BINNING pass and for the first tile in GMEM pass.