crashdec: Initial CP_MEMPOOL dumping
The mempool is, no surprise, a pool of memory that's used for 6
different queues. The queues are represented as a singly-linked list of
blocks, so that they can share memory dynamically. Most of this
information comes from the firmware mempool save/restore routine, which
use the same mechanism as the kernel dumper but actually parses the
linked lists in order to avoid saving/restoring all the blocks.

In addition, it turns out that these queues are more-or-less the
replacement for the old MEQ. Dumping this will let us see how far ahead
the SQE is from the rest of the GPU.