rnndec: handle UNK fields
To reduce noise, don't show unknown fields with zero value.  And if they
have a non-zero value, then show it in red.  This gets us the equivalent
benefit of not defining in xml valid bits of unknown purpose, without
loosing the benefit of documenting in the xml the fact that the bits
exist.