freedreno/rnn: Make rnn_decode_enum() respect variants
Sync from mesa commit 73241ca53e8e085b1e953ddb5afaa91c455eb038

We'll need this for afuc, and we're currently also open-coding the same
thing in rnnutils. It seems this function was added to decode pm4 packet
names, but it currently has no users, so make it useful for what it
was intended to do.

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/6140>