cmake -> meson
The goal is to use the same meson.build files in the subdirectories,
with our own toplevel meson.build

At this point, the subdir meson.build files are *similar* to what is
in mesa, the main difference is that we have yet to move the disasm
code to ir2/ir3.

There are a few hacks, including src/freedreno/registers symlink that
can be dropped after a small bit of refactoring on the mesa side.